Add test_c_specific
